Crop Production in Karamoja
The data visualization of the two crops in the various districts analyzed depicts that there is a need to improve the yield of the crops. The production of the crops per capita shows that without maximization of the area under cultivation the already low production will not be able to sustain the population in the near future. Hence additional technical input is required to enhance the yield so as meet the growing demand of the population.